+++ src/importlinter/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.8.0"
+__version__ = "1.9.0"
 
 from .application import output  # noqa
 from .domain import fields  # noqa