+++ tests/unit/domain/test_helpers.py
@@ -1,5 +1,5 @@
 import re
-from typing import List, Tuple, Optional
+from typing import List, Optional, Tuple
 
 import pytest
 from grimp import DetailedImport
@@ -325,6 +325,12 @@ class TestResolveImportExpressions:
             line_number=2,
             line_contents="from someotherpackage import two",
         ),
+        DirectImport(
+            importer=Module("mypackage.green.cats.very.deep.module.cats"),
+            imported=Module("mypackage.orange.mice.another.verydeep.one.dogs"),
+            line_number=1,
+            line_contents="-",
+        ),
     ]
 
     @pytest.mark.parametrize(
@@ -367,6 +373,34 @@ class TestResolveImportExpressions:
                     ImportExpression(importer="mypackage.*.cats", imported="mypackage.*.dogs"),
                 ],
                 set(DIRECT_IMPORTS[3:5]),
+            ),
+            (
+                "Importer recursive wildcard",
+                [
+                    ImportExpression(importer="mypackage.**", imported="mypackage.blue"),
+                ],
+                {DIRECT_IMPORTS[1]},
+            ),
+            (
+                "Imported recursive wildcard",
+                [
+                    ImportExpression(importer="mypackage.green", imported="mypackage.**"),
+                ],
+                set(DIRECT_IMPORTS[0:2]),
+            ),
+            (
+                "Importer and imported recursive wildcards",
+                [
+                    ImportExpression(importer="mypackage.**", imported="mypackage.**"),
+                ],
+                set(DIRECT_IMPORTS[0:6]) | {DIRECT_IMPORTS[8]},
+            ),
+            (
+                "Inner recursive wildcard",
+                [
+                    ImportExpression(importer="mypackage.**.cats", imported="mypackage.**.dogs"),
+                ],
+                set(DIRECT_IMPORTS[3:5]) | {DIRECT_IMPORTS[8]},
             ),
             (
                 "Multiple expressions, non-overlapping",