+++ CHANGELOG.rst
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+1.10.0 (Unreleased)
+-------------------
+
+* Recursive wildcard support for ignored imports.
+
+1.9.0 (2023-05-13)
+------------------
+
+* Update to Grimp 2.4.
+* Forbidden contracts: when include_external_packages is true, error if an external subpackage is
+  a forbidden module.
+
 1.8.0 (2023-03-03)
 ------------------